updated apis

+++ Famlib/myapp/views.py
@@ -338,48 +338,62 @@ def upload(request):
             return redirect(reverse('login'))
     return render(request, 'share.html')
 
-        
-@api_view(['GET', 'POST'])
-def users_list(request):
-    """
-    List all users, or create a new user.
-    """
-    if request.method == 'GET':
-        users = Users.objects.all()
-        serializer = UserSerializer(users, many=True)
-        return Response(serializer.data)
-
-    elif request.method == 'POST':
-        serializer = UserSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
-def users_detail(request, username):
+def users_detail(request, username=None,libraryid=None):
     """
     Retrieve, update or delete a user.
     """
-    try:
-        user = Users.objects.get(username=username)
-    except Users.DoesNotExist:
+    if libraryid != None:
+        libraryid = uuid.UUID(libraryid)
+        print(libraryid)
+        try:
+            data = Users.objects.filter(libraryid=libraryid)
+            serializer = UserSerializer(data, many=True)
+            return Response(serializer.data)
+        except Users.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+    elif username != None:
+        try:
+            user = Users.objects.get(username=username)
+        except Users.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == 'GET':
+            serializer = UserSerializer(user)
+            return Response(serializer.data)
+
+        elif request.method == 'PUT':
+            serializer = UserSerializer(user, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == 'DELETE':
+           user.delete()
+           return Response(status=status.HTTP_204_NO_CONTENT)
+    elif username==None and libraryid==None:
+        if request.method == 'GET':
+            users = Users.objects.all()
+            serializer = UserSerializer(users, many=True)
+            return Response(serializer.data)
+
+        elif request.method == 'POST':
+            serializer = UserSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    else:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == 'GET':
-        serializer = UserSerializer(user)
-        return Response(serializer.data)
 
-    elif request.method == 'PUT':
-        serializer = UserSerializer(user, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
     
 @api_view(['GET', 'POST'])
 def book_list(request, user_pk=None, library_name=None):
